Split out emit precedence handling. Don't use kOp_ style anymore, just use an array indexed by EmitOp.
