+++ source/slang/slang-c-like-source-emitter.h
@@ -8,6 +8,7 @@
 
 #include "slang-emit-context.h"
 #include "slang-extension-usage-tracker.h"
+#include "slang-emit-precedence.h"
 
 #include "ir.h"
 #include "ir-insts.h"
@@ -15,125 +16,6 @@
 
 namespace Slang
 {
-
-// Macros for setting up precedence
-#define SLANG_PRECEDENCE_LEFT(NAME)                      \
-    kEPrecedence_##NAME##_Left,     \
-    kEPrecedence_##NAME##_Right,
-
-#define SLANG_PRECEDENCE_RIGHT(NAME)                     \
-    kEPrecedence_##NAME##_Right,    \
-    kEPrecedence_##NAME##_Left,
-
-#define SLANG_PRECEDENCE_NON_ASSOC(NAME)                  \
-    kEPrecedence_##NAME##_Left,     \
-    kEPrecedence_##NAME##_Right = kEPrecedence_##NAME##_Left,
-
-#define SLANG_PRECEDENCE_EXPAND(NAME, ASSOC) SLANG_PRECEDENCE_##ASSOC(NAME)
-
-// x macro of precedence of types in order.
-// Used because in header, need to prefix macros to avoid clashes, and this style allows for prefixing without additional clutter
-#define SLANG_PRECEDENCE(x) \
-    x(None,         NON_ASSOC) \
-    x(Comma,        LEFT) \
-    \
-    x(General,      NON_ASSOC) \
-    \
-    x(Assign,       RIGHT) \
-    \
-    x(Conditional,  RIGHT) \
-    \
-    x(Or,           LEFT) \
-    x(And,          LEFT) \
-    x(BitOr,        LEFT) \
-    x(BitXor,       LEFT) \
-    x(BitAnd,       LEFT) \
-    \
-    x(Equality,     LEFT) \
-    x(Relational,   LEFT) \
-    x(Shift,        LEFT) \
-    x(Additive,     LEFT) \
-    x(Multiplicative, LEFT) \
-    x(Prefix,       RIGHT) \
-    x(Postfix,      LEFT) \
-    x(Atomic,       NON_ASSOC)
-
-// Precedence enum produced from the SLANG_PRECEDENCE macro
-enum EPrecedence
-{
-    SLANG_PRECEDENCE(SLANG_PRECEDENCE_EXPAND)
-};
-
-// Macro for define OpInfo and an associated enum type. Order or macro parameters is
-// Op, OpName, Precedence
-#define SLANG_OP_INFO(x) \
-    x(None, "", None) \
-    \
-    x(Comma, ",", Comma) \
-    \
-    x(General, "", General) \
-    \
-    x(Assign, "=", Assign) \
-    x(AddAssign, "+=", Assign) \
-    x(SubAssign, "-=", Assign) \
-    x(MulAssign, "*=", Assign) \
-    x(DivAssign, "/=", Assign) \
-    x(ModAssign, "%=", Assign) \
-    x(LshAssign, "<<=", Assign) \
-    x(RshAssign, ">>=", Assign) \
-    x(OrAssign, "|=", Assign) \
-    x(AndAssign, "&=", Assign) \
-    x(XorAssign, "^=", Assign) \
-    \
-    x(Conditional, "?:", Conditional) \
-    \
-    x(Or, "||", Or) \
-    x(And, "&&", And) \
-    x(BitOr, "|", BitOr) \
-    x(BitXor, "^", BitXor) \
-    x(BitAnd, "&", BitAnd) \
-    \
-    x(Eql, "==", Equality) \
-    x(Neq, "!=", Equality) \
-    \
-    x(Less, "<", Relational) \
-    x(Greater, ">", Relational) \
-    x(Leq, "<=", Relational) \
-    x(Geq, ">=", Relational) \
-    \
-    x(Lsh, "<<", Shift) \
-    x(Rsh, ">>", Shift) \
-    \
-    x(Add, "+", Additive) \
-    x(Sub, "-", Additive) \
-    \
-    x(Mul, "*", Multiplicative) \
-    x(Div, "/", Multiplicative) \
-    x(Mod, "%", Multiplicative) \
-    \
-    x(Prefix, "", Prefix) \
-    x(Postfix, "", Postfix) \
-    x(Atomic, "", Atomic)
-
-#define SLANG_OP_INFO_ENUM(op, name, precedence) op,
-
-enum class EmitOp
-{
-    SLANG_OP_INFO(SLANG_OP_INFO_ENUM)
-    CountOf,
-};
-
-// Info on an op for emit purposes
-struct EmitOpInfo
-{
-    SLANG_FORCE_INLINE static const EmitOpInfo& get(EmitOp inOp) { return s_infos[int(inOp)]; }
-
-    char const* op;
-    EPrecedence leftPrecedence;
-    EPrecedence rightPrecedence;
-
-    static const EmitOpInfo s_infos[int(EmitOp::CountOf)];
-};
 
 struct CLikeSourceEmitter
 {
@@ -249,9 +131,6 @@ struct CLikeSourceEmitter
     bool isTargetIntrinsicModifierApplicable(IRTargetIntrinsicDecoration* decoration);
 
     void emitStringLiteral(const String& value);
-
-    static EmitOpInfo leftSide(EmitOpInfo const& outerPrec, EmitOpInfo const& prec);
-    static EmitOpInfo rightSide(EmitOpInfo const& prec, EmitOpInfo const& outerPrec);
 
     void requireGLSLExtension(const String& name);
 
@@ -495,13 +374,6 @@ struct CLikeSourceEmitter
     EmitContext* m_context;
     SourceStream* m_stream;
 };
-
-// Precedence macros no longer needed
-#undef SLANG_PRECEDENCE_EXPAND
-#undef SLANG_PRECEDENCE_NON_ASSOC
-#undef SLANG_PRECEDENCE_RIGHT
-#undef SLANG_PRECEDENCE_LEFT
-
 
 }
 #endif
