+++ source/slang/slang-emit-context.cpp
@@ -1,4 +1,4 @@
-// slang-emit-emit.cpp
+// slang-emit-context.cpp
 #include "slang-emit-context.h"
 
 namespace Slang {
